Fix edge case for December (#11)
When translating `ucx-py`=>`rapids`, if the result would be December, it
would return zero for the month instead and the next year.

Example:
```
$ curl -sL https://version.gpuci.io/ucx-py/0.34
23.10

$ curl -sL https://version.gpuci.io/ucx-py/0.35
24.00

$ curl -sL https://version.gpuci.io/ucx-py/0.36
24.02
```

This PR adjusts for this edge case. I think there might be a way to
adjust the equations, but it felt cleaner to just add a check for this
edge case so that the math is clean too.